Some updated that I felt was important to do
I formated the document using Prettier.

Change to loadIllustrativeImage where the image is collected from an object with the correct file instead of a weird ternary.

Added Cors handle to the backend instead of using a proxy.

Added a 200 code respons to upload.

Changed the fetch-links to include server link

Removed unnecessary div.

Changed styling on upload component to absolute instead of fixed.

Changed onClick functions to the more efficient {function} instead of {() => function()}